separate property definition code from the team manager
